componentify the footer
